Finished most of the styleing that I wanted to do in both mobile and desktop view. All that left is to finished the Javascript component for the slideshow
This was the first time I added some animations to my CSS. It was a lot of fun and really helped my understand how some animations/ transitions work. What I hope to do next time is complete the slideshow component but I hope I can add some smooth transition to a background image in my header or else I'm going to have to give it a default look.